potential unity v1.5.3

+++ common/install.sh
@@ -5,7 +5,7 @@ fi
 
 cp -rf $INSTALLER/custom/$API/$ARCH/* $INSTALLER/system
 
-if $MAGISK; then 
+if $MAGISK && ! $SYSOVERRIDE; then 
   mv -f $INSTALLER/system/bin/app_process32_magisk $INSTALLER/system/bin/app_process32
   $IS64BIT && mv -f $INSTALLER/system/bin/app_process64_magisk $INSTALLER/system/bin/app_process64
   rm -f $INSTALLER/system/xposed.prop
